+++ djpcms/templatetags/tablesorter.py
@@ -4,5 +4,5 @@ from djpcms.views.changelist import ChangeList, table
 register = template.Library()
 
 @register.inclusion_tag('djpcms/tablesorter.html')
-def totable(queryset, djp):
-    return table(queryset, djp)
+def totable(queryset, djp, appmodel):
+    return table(queryset, djp, appmodel)